# Conflicts:
#	FakeNFT.xcodeproj/project.pbxproj
#	FakeNFT/Foundation/NetworkClient/NetworkClient.swift
#	FakeNFT/Resources/Assets.xcassets/Buttons/addCartButton.imageset/Contents.json
#	FakeNFT/Resources/Assets.xcassets/Buttons/heartEmpty.imageset/Contents.json
#	FakeNFT/Resources/Assets.xcassets/Buttons/heartFilled.imageset/Contents.json
#	FakeNFT/Resources/Assets.xcassets/Epic Profile/Contents.json
#	FakeNFT/Resources/Assets.xcassets/Epic Profile/userImageMock.imageset/Contents.json
#	FakeNFT/Resources/Assets.xcassets/NFT/Beige/BeigeApril.imageset/Contents.json
#	FakeNFT/Resources/Assets.xcassets/NFT/Beige/BeigeAurora.imageset/Contents.json
#	FakeNFT/Resources/Assets.xcassets/NFT/Beige/BeigeBimbo.imageset/Contents.json
#	FakeNFT/Resources/Assets.xcassets/NFT/Beige/BeigeBiscuit.imageset/Contents.json
#	FakeNFT/Resources/Assets.xcassets/NFT/Beige/Contents.json
#	FakeNFT/Resources/Assets.xcassets/Onboarding/Contents.json
#	FakeNFT/Resources/Assets.xcassets/Placeholders/Contents.json
#	FakeNFT/Resources/Assets.xcassets/Placeholders/approveDeleteNft.imageset/Contents.json
#	FakeNFT/Resources/Assets.xcassets/Tabs/Cart.imageset/Contents.json
#	FakeNFT/Resources/Assets.xcassets/Tabs/Profile.imageset/Contents.json
#	FakeNFT/Resources/Assets.xcassets/Tabs/Statistics.imageset/Contents.json
#	FakeNFT/Resources/Generated/Assets.swift
#	FakeNFT/Resources/Generated/Strings+Localizable.swift
#	FakeNFT/Resources/en.lproj/Localizable.strings
#	FakeNFT/Resources/ru.lproj/Localizable.strings
#	FakeNFT/SceneDelegate.swift
#	FakeNFT/Scenes/Common/Protocols/LoadingView.swift
#	FakeNFT/Scenes/Common/Views/LinePageControl.swift
#	FakeNFT/Scenes/TabBarController/TabBarController.swift

+++ FakeNFT/Scenes/Profile/ProfileMyNFT/MyNFTCell.swift
@@ -156,7 +156,7 @@ extension MyNFTCell {
         let name: String
         let rating: Int
         let author: String
-        let price: Float
+        let price: Double
         let isFavorite: Bool
         let id: String
     }
